+++ Editor/Drawers/DisabledDrawers/EnableIfAttributeDrawer.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using UnityEditor;
+using UnityEngine;
 
 namespace SaintsField.Editor.Drawers.DisabledDrawers
 {
@@ -10,20 +11,40 @@ namespace SaintsField.Editor.Drawers.DisabledDrawers
     public class EnableIfAttributeDrawer: ReadOnlyAttributeDrawer
     {
         protected override (string error, bool disabled) IsDisabled(SerializedProperty property,
-            ReadOnlyAttribute targetAttribute, FieldInfo info, Type type, object target)
+            ISaintsAttribute saintsAttribute, FieldInfo info, Type type, object target)
         {
+            EnableIfAttribute targetAttribute = (EnableIfAttribute) saintsAttribute;
+
+            EMode editorMode = targetAttribute.EditorMode;
+            bool editorRequiresEdit = editorMode.HasFlag(EMode.Edit);
+            bool editorRequiresPlay = editorMode.HasFlag(EMode.Play);
+
+            bool editorModeIsTrue = (
+                !editorMode.HasFlag(EMode.Edit) || !EditorApplication.isPlaying
+            ) && (
+                !editorMode.HasFlag(EMode.Play) || EditorApplication.isPlaying
+            );
+
             string[] bys = targetAttribute.ReadOnlyBys;
-            if(bys is null)
+            if (bys == null && editorRequiresEdit && editorRequiresPlay)
             {
-                return ("", targetAttribute.readOnlyDirectValue);
+#if SAINTSFIELD_DEBUG && SAINTSFIELD_DEBUG_READ_ONLY
+                Debug.Log($"return={editorModeIsTrue}");
+#endif
+                return ("", true);
             }
 
             List<bool> callbackTruly = new List<bool>();
+            if(!(editorRequiresEdit && editorRequiresPlay))
+            {
+                callbackTruly.Add(editorModeIsTrue);
+            }
+
             List<string> errors = new List<string>();
 
-            foreach (string andCallback in bys)
+            foreach (string orCallback in bys ?? Array.Empty<string>())
             {
-                (string error, bool isTruly) = IsTruly(target, type, andCallback);
+                (string error, bool isTruly) = IsTruly(target, type, orCallback);
                 if (error != "")
                 {
                     errors.Add(error);
@@ -33,13 +54,7 @@ namespace SaintsField.Editor.Drawers.DisabledDrawers
 
             if (errors.Count > 0)
             {
-                return (string.Join("\n\n", errors), true);
-            }
-
-            // empty means hide
-            if (callbackTruly.Count == 0)
-            {
-                return ("", targetAttribute.readOnlyDirectValue);
+                return (string.Join("\n\n", errors), false);
             }
 
             // or, get enabled