+++ Editor/Drawers/VisibilityDrawers/ShowIfAttributeDrawer.cs
@@ -9,15 +9,30 @@ namespace SaintsField.Editor.Drawers.VisibilityDrawers
     [CustomPropertyDrawer(typeof(ShowIfAttribute))]
     public class ShowIfAttributeDrawer: VisibilityAttributeDrawer
     {
+        // and
         protected override (string error, bool shown) IsShown(SerializedProperty property, ISaintsAttribute saintsAttribute, FieldInfo info,
             Type type, object target)
         {
             ShowIfAttribute showIfAttribute = (ShowIfAttribute)saintsAttribute;
 
+            EMode editorMode = showIfAttribute.EditorMode;
+            bool editorRequiresEdit = editorMode.HasFlag(EMode.Edit);
+            bool editorRequiresPlay = editorMode.HasFlag(EMode.Play);
+
+            bool editorModeIsTrue = (
+                !editorRequiresEdit || !EditorApplication.isPlaying
+            ) && (
+                !editorRequiresPlay || EditorApplication.isPlaying
+            );
+
             List<bool> callbackTruly = new List<bool>();
             List<string> errors = new List<string>();
+            if(!(editorRequiresEdit && editorRequiresPlay))
+            {
+                callbackTruly.Add(editorModeIsTrue);
+            }
 
-            foreach (string andCallback in showIfAttribute.orCallbacks)
+            foreach (string andCallback in showIfAttribute.Callbacks)
             {
                 (string error, bool isTruly) = IsTruly(target, type, andCallback);
                 if (error != "")
@@ -32,12 +47,7 @@ namespace SaintsField.Editor.Drawers.VisibilityDrawers
                 return (string.Join("\n\n", errors), true);
             }
 
-            // empty means show
-            if(callbackTruly.Count == 0)
-            {
-                return ("", true);
-            }
-
+            // empty = true
             bool truly = callbackTruly.All(each => each);
 
             return ("", truly);