+++ Runtime/DisableIfAttribute.cs
@@ -5,7 +5,11 @@ namespace SaintsField
     [AttributeUsage(AttributeTargets.Field, AllowMultiple = true)]
     public class DisableIfAttribute: ReadOnlyAttribute
     {
-        public DisableIfAttribute(bool directValue = true, string groupBy = "") : base(directValue, groupBy)
+        // public DisableIfAttribute(EMode editorMode, string groupBy = "") : base(editorMode, groupBy)
+        // {
+        // }
+
+        public DisableIfAttribute(EMode editorMode, params string[] by) : base(editorMode, by)
         {
         }