+++ Runtime/EMode.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace SaintsField
+{
+    [Flags]
+    public enum EMode
+    {
+        Edit = 1,
+        Play = 1 << 1,
+    }
+}